move AppNav component to its own folder. Remove useless test.
