--- src/components/App.test.js
@@ -1,4 +1,0 @@
-// This is a placeholder.
-// TODO: write better tests for components
-
-test("return undefined", () => undefined);
