+++ src/components/Layout.js
@@ -1,6 +1,6 @@
 import { Container } from "react-bootstrap";
 import AppFooter from "./AppFooter/AppFooter";
-import AppNav from "./AppNav";
+import AppNav from "./AppNav/AppNav";
 
 const Layout = ({ children }) => {
   return (
